requirements.txt
streamlit
pandas
numpy
tensorflow-cpu==2.7.0